pyproject.toml
[project]
name = "topological-wind-engineering"
version = "0.1.0"
description = "TDA tools for wind energy: persistent homology, Mapper, path homology for monitoring and prediction"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scipy>=1.7.0",
    "scikit-learn>=1.0.0",
    "matplotlib>=3.4.0",
    "ripser>=0.6.0",
    "requests>=2.26.0",
    "persim>=0.3.0",
    "torch>=1.9.0",
]

[project.optional-dependencies]
nrel = [
    "openoa[examples]>=2.2.0",
]

